refactor authorization files

+++ api/user/session.ts
@@ -1,5 +1,3 @@
-/* eslint @typescript-eslint/no-explicit-any: off */
-/* eslint camelcase: off */
 import { Request, Response } from "express";
 import { verifyToken } from "node-2fa";
 import UAParser from "ua-parser-js";
@@ -8,16 +6,19 @@ import User from "../../src/auth/User";
 import getAuthorization from "../../src/auth/getAuthorization";
 import { query } from "../../src/mysql";
 import hash from "../../src/util/hash";
+import rejectRequest from "../../src/util/rejectRequest";
 import snowflake from "../../src/util/snowflake";
 
 // FIXME - Clean up this file
 
 export const route = "auth/session";
+export default async function api(req: Request, res: Response) {
 
-export default async function api(req: Request, res: Response): Promise<any> {
-
-	// Parse the request
-	const body = { ...req.body, ...req.query };
+	// Get body
+	const body: Record<string, string | undefined> = { ...req.body, ...req.query };
+	
+	// Check method
+	if ([ "POST", "DELETE", "GET" ].indexOf(req.method) === -1) return rejectRequest(res, 405, `Method '${ req.method }' not allowed.`);
 
 	// Make sure method is POST
 	if (req.method === "POST") {
@@ -26,30 +27,15 @@ export default async function api(req: Request, res: Response): Promise<any> {
 		const { email, password, token, noredirect } = body;
 
 		// Ensure Fields are there
-		const requiredFields = [ "email", "password" ];
-		requiredFields.map(field => {
-			if (body[field] === undefined) return res.status(406).json({
-				success: false,
-				error: "406 Not Acceptable",
-				description: `Field '${ field }' is required but received 'undefined'.`,
-				readable: `Please enter a ${ field }.`
-			});
-		});
-
-		// Continue?
-		if (res.headersSent) return;
+		if (!email) return rejectRequest(res, 406, "Required field 'email' is missing.");
+		if (!password) return rejectRequest(res, 406, "Required field 'password' is missing.");
 
 		// Generate password hash
 		const md5 = hash(password);
 
 		// Select users with the same email address and password
 		const users = await query<MySQLData.User>(`SELECT * FROM users WHERE email = "${ email }" AND passwd_md5 = "${ md5 }";`);
-		if (users.length === 0) return res.status(406).json({
-			success: false,
-			error: "406 Not Acceptable",
-			description: "Fields 'email' or 'password' are invalid.",
-			readable: "Incorrect password."
-		});
+		if (users.length === 0) return rejectRequest(res, 401, "Incorrect email or password.");
 
 		// Get user from database
 		const [ user ] = users;
@@ -59,45 +45,32 @@ export default async function api(req: Request, res: Response): Promise<any> {
 		if (mfa !== undefined && mfa.pending === 0) {
 
 			// If no token
-			if (token === undefined || token.length === 0) return res.status(417).json({
-				success: false,
-				message: "417 Expectation Failed",
-				description: "This account requires further authentication."
-			});
+			if (token === undefined || token.length === 0) return rejectRequest(res, 406, "Required field 'token' is missing.");
 
 			// Verify token
 			const verify = verifyToken(mfa.secret, token);
-			if (verify === null || verify.delta !== 0) return res.status(406).json({
-				success: false,
-				message: "406 Not Acceptable",
-				description: "The token is not correct.",
-				readable: "Invalid token."
-			});
+			if (verify === null || verify.delta !== 0) return rejectRequest(res, 401, "Invalid 2FA token.");
 
 		}
 
 		// Generate session id
-		const session_id = uuid();
-
 		const now = Date.now();
+		const session_id = uuid();
 
 		// Insert into sessions
 		await query(`INSERT INTO sessions (id, session_id, user, md5, created_ms, last_used_ms, user_agent, ip_address) VALUES (${ snowflake() }, "${ session_id }", ${ user.id }, "${ md5 }", ${ now }, ${ now }, "${ req.header("User-Agent") }", "${ req.ip }");`);
 
-		// Set cookie
+		// Set session
 		res.cookie("session_id", session_id, { maxAge: 1000 * 60 * 60 * 24 * (body.hasOwnProperty("rememberme") ? 3650 : 7) });
 		res.header("authorization", session_id);
 
 		// Respond with redirect to generate session
-		if (noredirect) {
-			res.json({
-				success: true,
-				session_id
-			});
-			return;
-		}
-		res.redirect(307, "./@me");
-		return;
+		if (noredirect) return res.json({
+			success: true,
+			session_id
+		});
+		
+		return res.redirect(307, "./@me");
 
 	}
 
@@ -105,67 +78,31 @@ export default async function api(req: Request, res: Response): Promise<any> {
 	if (req.method === "DELETE") {
 
 		// Get fields
-		const session_id = body.session_id || req.cookies.session_id;
-
-		// Ensure Fields are there
-		if (session_id === undefined) return res.status(406).json({
-			success: false,
-			error: "406 Not Acceptable",
-			description: "Field 'session_id' is required but received 'undefined'."
-		});
-
-		// Continue?
-		if (res.headersSent) return;
+		const authorization = body.authorization || body.session_id || getAuthorization(req);
+		if (!authorization) return rejectRequest(res, 406, "Required field 'authorization' is missing.");
 
 		// Get current session
-		const [ session ] = await query<MySQLData.Session>(`SELECT * FROM sessions WHERE session_id = "${ session_id }";`);
-
-		if (session === undefined) return res.status(406).json({
-			success: false,
-			error: "406 Not Acceptable",
-			description: `Session ID '${ session_id }' is not a valid session ID.`
-		});
-
-		if (res.headersSent) return;
+		const [ session ] = await query<MySQLData.Session>(`SELECT * FROM sessions WHERE session_id = "${ authorization }";`);
+		if (session === undefined) return rejectRequest(res, 401, "Invalid session ID.");
 
 		// Delete from sessions
-		if (body.all === true) {
-			await query(`DELETE FROM sessions WHERE user = ${ session.user }`);
-		} else {
-			await query(`DELETE FROM sessions WHERE session_id = "${ session_id }"`);
-		}
+		await query(`DELETE FROM sessions WHERE session_id = "${ authorization }"`);
+		if ("all" in body) await query(`DELETE FROM sessions WHERE user = ${ session.user }`);
 
 		// Delete cookie
-		if (session_id === getAuthorization(req)) res.cookie("session_id", "", { maxAge: 0 });
+		if (authorization === getAuthorization(req)) res.cookie("session_id", "", { maxAge: 0 });
 
 		// Respond with session
-		return res.json({
-			success: true
-		});
+		return res.json({ success: true });
 
 	}
 
 	// If trying to list all sessions
 	if (req.method === "GET") {
 
-		// Verify authorization
 		const authorization = getAuthorization(req);
-		if (authorization === undefined) return res.status(401).json({
-			success: false,
-			message: "401 Unauthorized",
-			description: "You likley do not have a valid session token."
-		});
-
-		// Get user and 2fa status
-		const user = await User.fromAuthorization(authorization);
-		if (!user) return res.status(401).json({
-			success: false,
-			message: "401 Unauthorized",
-			description: "You likley do not have a valid session token."
-		});
-
-		// Ensure getUser didnt reject the request
-		if (res.headersSent) return;
+		const user = authorization && await User.fromAuthorization(authorization);
+		if (!authorization || !user) return rejectRequest(res, 401);
 
 		// Get sessions
 		let sessions = await query(`SELECT * FROM sessions WHERE user = ${ user.id }`);
@@ -178,16 +115,11 @@ export default async function api(req: Request, res: Response): Promise<any> {
 		});
 
 		// Respond with sessions
-		res.json({ sessions });
-		return;
+		return res.json({
+			success: true,
+			sessions
+		});
 
 	}
-
-	// Respond with 405
-	return res.status(405).json({
-		success: false,
-		error: "405 Method Not Allowed",
-		description: `Method '${ req.method }' is not allowed on this endpoint.`
-	});
 
 }
